Rename badge-status component
