+++ src/components/badge-status/badge-status.scss
