+++ src/components/pull-request-line/pull-request-line.js
@@ -1,4 +1,4 @@
-import Badge from '../badge/badge.vue'
+import Badge from '@/components/badge-status/badge-status.vue'
 import LivingIcon from '../living-icon/living-icon.vue'
 import UpdateIcon from '../update-icon/update-icon.vue'
 import BuildStatuses from '../build-statuses/build-statuses.vue'
